requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
deep-translator==1.11.4
Django==5.1.7
django-redis==5.4.0
idna==3.10
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
